+++ source/ada/lsp-ada_handlers-scenario_variables_commands.ads
@@ -1,0 +1,48 @@
+------------------------------------------------------------------------------
+--                         Language Server Protocol                         --
+--                                                                          --
+--                     Copyright (C) 2026, AdaCore                         --
+--                                                                          --
+-- This is free software;  you can redistribute it  and/or modify it  under --
+-- terms of the  GNU General Public License as published  by the Free Soft- --
+-- ware  Foundation;  either version 3,  or (at your option) any later ver- --
+-- sion.  This software is distributed in the hope  that it will be useful, --
+-- but WITHOUT ANY WARRANTY;  without even the implied warranty of MERCHAN- --
+-- TABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public --
+-- License for  more details.  You should have  received  a copy of the GNU --
+-- General  Public  License  distributed  with  this  software;   see  file --
+-- COPYING3.  If not, go to http://www.gnu.org/licenses for a complete copy --
+-- of the license.                                                          --
+------------------------------------------------------------------------------
+
+--  Implementation of the command to retrieve scenario (external) variables
+--  information for the loaded project tree, using GPR2's
+--  GPR2.Project.External API. This is used by the client to populate the
+--  Scenario View.
+
+with LSP.Ada_Commands;
+with LSP.Errors;
+with LSP.Server_Jobs;
+
+package LSP.Ada_Handlers.Scenario_Variables_Commands is
+
+   type Command is new LSP.Ada_Commands.Command with null record;
+
+   overriding
+   function Create
+     (Any : not null access LSP.Structures.LSPAny_Vector) return Command;
+
+   overriding
+   procedure Execute
+     (Self     : Command;
+      Handler  : not null access LSP.Ada_Handlers.Message_Handler'Class;
+      Response : in out LSP.Structures.LSPAny_Or_Null;
+      Error    : in out LSP.Errors.ResponseError_Optional);
+
+   overriding
+   function Priority (Self : Command) return LSP.Server_Jobs.Job_Priority
+   is (LSP.Server_Jobs.High);
+
+   for Command'External_Tag use "als-scenario-variables-information";
+
+end LSP.Ada_Handlers.Scenario_Variables_Commands;